+++ tests/test_cc_pcan_can.py
@@ -16,6 +16,7 @@ import can as python_can
 import pytest
 
 from pykiso import Message
+from pykiso.lib.connectors import cc_pcan_can
 from pykiso.lib.connectors.cc_pcan_can import CCPCanCan, PCANBasic, can
 from pykiso.message import (
     MessageAckType,
@@ -115,6 +116,7 @@ def mock_PCANBasic(mocker):
                 "remote_id": None,
                 "can_filters": None,
                 "logging_activated": True,
+                "bus_error_warning_filter": False,
             },
         ),
         (
@@ -140,6 +142,7 @@ def mock_PCANBasic(mocker):
                     {"can_id": 0x507, "can_mask": 0x7FF, "extended": False}
                 ],
                 "logging_activated": False,
+                "bus_error_warning_filter": True,
             },
             {
                 "interface": "pcan",
@@ -163,6 +166,7 @@ def mock_PCANBasic(mocker):
                     {"can_id": 0x507, "can_mask": 0x7FF, "extended": False}
                 ],
                 "logging_activated": False,
+                "bus_error_warning_filter": True,
             },
         ),
     ],
@@ -170,9 +174,11 @@ def mock_PCANBasic(mocker):
 def test_constructor(constructor_params, expected_config, caplog):
 
     param = constructor_params.values()
+    log = logging.getLogger("can.pcan")
+
     with caplog.at_level(logging.WARNING):
         can_inst = CCPCanCan(*param)
-
+    log.warning("Bus error: an error counter")
     assert can_inst.interface == expected_config["interface"]
     assert can_inst.channel == expected_config["channel"]
     assert can_inst.bitrate == expected_config["bitrate"]
@@ -196,6 +202,11 @@ def test_constructor(constructor_params, expected_config, caplog):
 
     if not can_inst.is_fd and can_inst.enable_brs:
         assert "Bitrate switch will have no effect" in caplog.text
+
+    if expected_config["bus_error_warning_filter"]:
+        assert "Bus error: an error counter" not in caplog.text
+    else:
+        assert "Bus error: an error counter" in caplog.text
 
 
 @pytest.mark.parametrize(